Capstone V1/Azure Function API/my-function-app/requirements.txt
#Author: Mackenzie Van Vliet

# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
flask
flask-cors
torch
torchvision
numpy

